Move viewsets to views.py and remove viewsets.py
