+++ nautobot_device_lifecycle_mgmt/views.py
@@ -10,32 +10,15 @@ from django.conf import settings
 from django.db.models import Count, ExpressionWrapper, F, FloatField, Q
 from django_tables2 import RequestConfig
 from matplotlib.ticker import MaxNLocator
+from nautobot.apps.views import NautobotUIViewSet
+from nautobot.core.forms.search import SearchForm
 from nautobot.core.views import generic
 from nautobot.core.views.mixins import ContentTypePermissionRequiredMixin
 from nautobot.core.views.paginator import EnhancedPaginator, get_paginate_count
+from nautobot.dcim.models import Device
 
-from nautobot_device_lifecycle_mgmt import choices
-from nautobot_device_lifecycle_mgmt.filters import (
-    DeviceSoftwareValidationResultFilterSet,
-    InventoryItemSoftwareValidationResultFilterSet,
-)
-from nautobot_device_lifecycle_mgmt.forms import (
-    DeviceSoftwareValidationResultFilterForm,
-    InventoryItemSoftwareValidationResultFilterForm,
-)
-from nautobot_device_lifecycle_mgmt.models import (
-    DeviceSoftwareValidationResult,
-    InventoryItemSoftwareValidationResult,
-    SoftwareImageLCM,
-    SoftwareLCM,
-)
-from nautobot_device_lifecycle_mgmt.tables import (
-    DeviceSoftwareValidationResultListTable,
-    DeviceSoftwareValidationResultTable,
-    InventoryItemSoftwareValidationResultListTable,
-    InventoryItemSoftwareValidationResultTable,
-    SoftwareImageLCMTable,
-)
+from nautobot_device_lifecycle_mgmt import choices, filters, forms, models, tables
+from nautobot_device_lifecycle_mgmt.api import serializers
 from nautobot_device_lifecycle_mgmt.utils import count_related_m2m
 
 PLUGIN_CFG = settings.PLUGINS_CONFIG["nautobot_device_lifecycle_mgmt"]
@@ -48,21 +31,186 @@ logger = logging.getLogger("nautobot_device_lifecycle_mgmt")
 GREEN, RED, GREY = ("#D5E8D4", "#F8CECC", "#808080")
 
 
+class HardwareLCMUIViewSet(NautobotUIViewSet):
+    """HardwareLCM UI ViewSet."""
+
+    bulk_update_form_class = forms.HardwareLCMBulkEditForm
+    filterset_class = filters.HardwareLCMFilterSet
+    filterset_form_class = forms.HardwareLCMFilterForm
+    form_class = forms.HardwareLCMForm
+    queryset = models.HardwareLCM.objects.prefetch_related("device_type")
+    serializer_class = serializers.HardwareLCMSerializer
+    table_class = tables.HardwareLCMTable
+
+    def get_extra_context(self, request, instance):  # pylint: disable=signature-differs
+        """Return any additional context data for the template.
+
+        request: The current request
+        instance: The object being viewed
+        """
+        if not instance:
+            return {}
+        if instance.device_type:
+            return {"devices": Device.objects.restrict(request.user, "view").filter(device_type=instance.device_type)}
+        if instance.inventory_item:
+            return {
+                "devices": Device.objects.restrict(request.user, "view").filter(
+                    inventory_items__part_id=instance.inventory_item
+                )
+            }
+        return {"devices": []}
+
+
+class SoftwareLCMUIViewSet(NautobotUIViewSet):
+    """SoftwareLCM UI ViewSet."""
+
+    # TODO: Add bulk edit form
+    # bulk_update_form_class = forms.SoftwareLCMBulkEditForm
+    filterset_class = filters.SoftwareLCMFilterSet
+    filterset_form_class = forms.SoftwareLCMFilterForm
+    form_class = forms.SoftwareLCMForm
+    queryset = models.SoftwareLCM.objects.prefetch_related("device_platform")
+    serializer_class = serializers.SoftwareLCMSerializer
+    table_class = tables.SoftwareLCMTable
+
+    def get_extra_context(self, request, instance):  # pylint: disable=signature-differs
+        """Changes "Softwares" => "Software"."""
+        search_form = SearchForm(data=self.request.GET)
+
+        return {
+            "search_form": search_form,
+            "title": "Software",
+            "verbose_name_plural": "Software",
+        }
+
+
+class SoftwareImageLCMUIViewSet(NautobotUIViewSet):
+    """SoftwareImageLCM UI ViewSet."""
+
+    # TODO: Add bulk edit form
+    # bulk_update_form_class = forms.SoftwareImageLCMBulkEditForm
+    filterset_class = filters.SoftwareImageLCMFilterSet
+    filterset_form_class = forms.SoftwareImageLCMFilterForm
+    form_class = forms.SoftwareImageLCMForm
+    queryset = models.SoftwareImageLCM.objects.all()
+    serializer_class = serializers.SoftwareImageLCMSerializer
+    table_class = tables.SoftwareImageLCMTable
+
+
+class ValidatedSoftwareLCMUIViewSet(NautobotUIViewSet):
+    """ValidatedSoftwareLCM UI ViewSet."""
+
+    # TODO: Add bulk edit form
+    # bulk_update_form_class = forms.ValidatedSoftwareLCMBulkEditForm
+    filterset_class = filters.ValidatedSoftwareLCMFilterSet
+    filterset_form_class = forms.ValidatedSoftwareLCMFilterForm
+    form_class = forms.ValidatedSoftwareLCMForm
+    queryset = models.ValidatedSoftwareLCM.objects.all()
+    serializer_class = serializers.ValidatedSoftwareLCMSerializer
+    table_class = tables.ValidatedSoftwareLCMTable
+
+
+class ContractLCMUIViewSet(NautobotUIViewSet):
+    """ContractLCM UI ViewSet."""
+
+    bulk_update_form_class = forms.ContractLCMBulkEditForm
+    filterset_class = filters.ContractLCMFilterSet
+    filterset_form_class = forms.ContractLCMFilterForm
+    form_class = forms.ContractLCMForm
+    queryset = models.ContractLCM.objects.all()
+    serializer_class = serializers.ContractLCMSerializer
+    table_class = tables.ContractLCMTable
+
+    def get_extra_context(self, request, instance):  # pylint: disable=signature-differs
+        """Return any additional context data for the template.
+
+        request: The current request
+        instance: The object being viewed
+        """
+        return {
+            "contacts": models.ContactLCM.objects.restrict(request.user, "view")
+            .filter(contract=instance)
+            .exclude(type="Owner")
+            .order_by("type", "priority"),
+            "owners": models.ContactLCM.objects.restrict(request.user, "view")
+            .filter(contract=instance, type="Owner")
+            .order_by("type", "priority"),
+        }
+
+
+class ProviderLCMUIViewSet(NautobotUIViewSet):
+    """ProviderLCM UI ViewSet."""
+
+    bulk_update_form_class = forms.ProviderLCMBulkEditForm
+    filterset_class = filters.ProviderLCMFilterSet
+    filterset_form_class = forms.ProviderLCMFilterForm
+    form_class = forms.ProviderLCMForm
+    queryset = models.ProviderLCM.objects.all()
+    serializer_class = serializers.ProviderLCMSerializer
+    table_class = tables.ProviderLCMTable
+
+    def get_extra_context(self, request, instance):  # pylint: disable=signature-differs
+        """Return any additional context data for the template.
+
+        request: The current request
+        instance: The object being viewed
+        """
+        return {"contracts": models.ContractLCM.objects.restrict(request.user, "view").filter(provider=instance)}
+
+
+class ContactLCMUIViewSet(NautobotUIViewSet):
+    """ContactLCM UI ViewSet."""
+
+    bulk_update_form_class = forms.ContactLCMBulkEditForm
+    filterset_class = filters.ContactLCMFilterSet
+    filterset_form_class = forms.ContactLCMFilterForm
+    form_class = forms.ContactLCMForm
+    queryset = models.ContactLCM.objects.all()
+    serializer_class = serializers.ContactLCMSerializer
+    table_class = tables.ContactLCMTable
+
+
+class CVELCMUIViewSet(NautobotUIViewSet):
+    """CVELCM UI ViewSet."""
+
+    bulk_update_form_class = forms.CVELCMBulkEditForm
+    filterset_class = filters.CVELCMFilterSet
+    filterset_form_class = forms.CVELCMFilterForm
+    form_class = forms.CVELCMForm
+    queryset = models.CVELCM.objects.all()
+    serializer_class = serializers.CVELCMSerializer
+    table_class = tables.CVELCMTable
+
+
+class VulnerabilityLCMUIViewSet(NautobotUIViewSet):
+    """VulnerabilityLCM UI ViewSet."""
+
+    bulk_update_form_class = forms.VulnerabilityLCMBulkEditForm
+    filterset_class = filters.VulnerabilityLCMFilterSet
+    filterset_form_class = forms.VulnerabilityLCMFilterForm
+    form_class = forms.VulnerabilityLCMForm
+    queryset = models.VulnerabilityLCM.objects.all()
+    serializer_class = serializers.VulnerabilityLCMSerializer
+    table_class = tables.VulnerabilityLCMTable
+
+
 class SoftwareSoftwareImagesLCMView(generic.ObjectView):
     """Software Images tab for Software view."""
 
-    queryset = SoftwareLCM.objects.all()
+    queryset = models.SoftwareLCM.objects.all()
     template_name = "nautobot_device_lifecycle_mgmt/softwarelcm_software_images.html"
 
     def get_extra_context(self, request, instance):
         """Adds Software Images table."""
         softwareimages = (
-            instance.software_images.annotate(device_type_count=count_related_m2m(SoftwareImageLCM, "device_types"))
-            .annotate(object_tag_count=count_related_m2m(SoftwareImageLCM, "object_tags"))
+            instance.software_images.annotate(
+                device_type_count=count_related_m2m(models.SoftwareImageLCM, "device_types")
+            )
+            .annotate(object_tag_count=count_related_m2m(models.SoftwareImageLCM, "object_tags"))
             .restrict(request.user, "view")
         )
 
-        softwareimages_table = SoftwareImageLCMTable(data=softwareimages, user=request.user, orderable=False)
+        softwareimages_table = tables.SoftwareImageLCMTable(data=softwareimages, user=request.user, orderable=False)
 
         paginate = {
             "paginator_class": EnhancedPaginator,
@@ -207,12 +355,12 @@ class ReportOverviewHelper(ContentTypePermissionRequiredMixin, generic.View):
 class ValidatedSoftwareDeviceReportView(generic.ObjectListView):
     """View for executive report on software Validation."""
 
-    filterset = DeviceSoftwareValidationResultFilterSet
-    filterset_form = DeviceSoftwareValidationResultFilterForm
-    table = DeviceSoftwareValidationResultTable
+    filterset = filters.DeviceSoftwareValidationResultFilterSet
+    filterset_form = forms.DeviceSoftwareValidationResultFilterForm
+    table = tables.DeviceSoftwareValidationResultTable
     template_name = "nautobot_device_lifecycle_mgmt/validatedsoftware_device_report.html"
     queryset = (
-        DeviceSoftwareValidationResult.objects.values("device__device_type__model", "device__device_type__pk")
+        models.DeviceSoftwareValidationResult.objects.values("device__device_type__model", "device__device_type__pk")
         .distinct()
         .annotate(
             total=Count("device__device_type__model"),
@@ -232,16 +380,18 @@ class ValidatedSoftwareDeviceReportView(generic.ObjectListView):
         super().setup(request, *args, **kwargs)  #
         try:
             report_last_run = (
-                DeviceSoftwareValidationResult.objects.filter(run_type=choices.ReportRunTypeChoices.REPORT_FULL_RUN)
+                models.DeviceSoftwareValidationResult.objects.filter(
+                    run_type=choices.ReportRunTypeChoices.REPORT_FULL_RUN
+                )
                 .latest("last_updated")
                 .last_run
             )
-        except DeviceSoftwareValidationResult.DoesNotExist:  # pylint: disable=no-member
+        except models.DeviceSoftwareValidationResult.DoesNotExist:  # pylint: disable=no-member
             report_last_run = None
 
         device_aggr = self.get_global_aggr(request)
         _platform_qs = (
-            DeviceSoftwareValidationResult.objects.values("device__platform__name")
+            models.DeviceSoftwareValidationResult.objects.values("device__platform__name")
             .distinct()
             .annotate(
                 total=Count("device__platform__name"),
@@ -279,7 +429,7 @@ class ValidatedSoftwareDeviceReportView(generic.ObjectListView):
         Returns:
             device_aggr: device global report dict
         """
-        device_qs = DeviceSoftwareValidationResult.objects
+        device_qs = models.DeviceSoftwareValidationResult.objects
 
         device_aggr = {}
         if self.filterset is not None:
@@ -339,10 +489,10 @@ class ValidatedSoftwareDeviceReportView(generic.ObjectListView):
 class DeviceSoftwareValidationResultListView(generic.ObjectListView):
     """DeviceSoftawareValidationResult List view."""
 
-    queryset = DeviceSoftwareValidationResult.objects.all()
-    filterset = DeviceSoftwareValidationResultFilterSet
-    filterset_form = DeviceSoftwareValidationResultFilterForm
-    table = DeviceSoftwareValidationResultListTable
+    queryset = models.DeviceSoftwareValidationResult.objects.all()
+    filterset = filters.DeviceSoftwareValidationResultFilterSet
+    filterset_form = forms.DeviceSoftwareValidationResultFilterForm
+    table = tables.DeviceSoftwareValidationResultListTable
     action_buttons = ("export",)
     template_name = "nautobot_device_lifecycle_mgmt/devicesoftwarevalidationresult_list.html"
 
@@ -350,12 +500,12 @@ class DeviceSoftwareValidationResultListView(generic.ObjectListView):
 class ValidatedSoftwareInventoryItemReportView(generic.ObjectListView):
     """View for executive report on inventory item software validation."""
 
-    filterset = InventoryItemSoftwareValidationResultFilterSet
-    filterset_form = InventoryItemSoftwareValidationResultFilterForm
-    table = InventoryItemSoftwareValidationResultTable
+    filterset = filters.InventoryItemSoftwareValidationResultFilterSet
+    filterset_form = forms.InventoryItemSoftwareValidationResultFilterForm
+    table = tables.InventoryItemSoftwareValidationResultTable
     template_name = "nautobot_device_lifecycle_mgmt/validatedsoftware_inventoryitem_report.html"
     queryset = (
-        InventoryItemSoftwareValidationResult.objects.values(
+        models.InventoryItemSoftwareValidationResult.objects.values(
             "inventory_item__part_id",
             "inventory_item__name",
             "inventory_item__pk",
@@ -381,18 +531,18 @@ class ValidatedSoftwareInventoryItemReportView(generic.ObjectListView):
         super().setup(request, *args, **kwargs)
         try:
             report_last_run = (
-                InventoryItemSoftwareValidationResult.objects.filter(
+                models.InventoryItemSoftwareValidationResult.objects.filter(
                     run_type=choices.ReportRunTypeChoices.REPORT_FULL_RUN
                 )
                 .latest("last_updated")
                 .last_run
             )
-        except InventoryItemSoftwareValidationResult.DoesNotExist:  # pylint: disable=no-member
+        except models.InventoryItemSoftwareValidationResult.DoesNotExist:  # pylint: disable=no-member
             report_last_run = None
 
         inventory_aggr = self.get_global_aggr(request)
         _platform_qs = (
-            InventoryItemSoftwareValidationResult.objects.values("inventory_item__manufacturer__name")
+            models.InventoryItemSoftwareValidationResult.objects.values("inventory_item__manufacturer__name")
             .distinct()
             .annotate(
                 total=Count("inventory_item__manufacturer__name"),
@@ -432,7 +582,7 @@ class ValidatedSoftwareInventoryItemReportView(generic.ObjectListView):
         Returns:
             inventory_aggr: inventory item global report dict
         """
-        inventory_item_qs = InventoryItemSoftwareValidationResult.objects
+        inventory_item_qs = models.InventoryItemSoftwareValidationResult.objects
 
         inventory_aggr = {}
         if self.filterset is not None:
@@ -499,9 +649,9 @@ class ValidatedSoftwareInventoryItemReportView(generic.ObjectListView):
 class InventoryItemSoftwareValidationResultListView(generic.ObjectListView):
     """InvenotryItemSoftawareValidationResult List view."""
 
-    queryset = InventoryItemSoftwareValidationResult.objects.all()
-    filterset = InventoryItemSoftwareValidationResultFilterSet
-    filterset_form = InventoryItemSoftwareValidationResultFilterForm
-    table = InventoryItemSoftwareValidationResultListTable
+    queryset = models.InventoryItemSoftwareValidationResult.objects.all()
+    filterset = filters.InventoryItemSoftwareValidationResultFilterSet
+    filterset_form = forms.InventoryItemSoftwareValidationResultFilterForm
+    table = tables.InventoryItemSoftwareValidationResultListTable
     action_buttons = ("export",)
     template_name = "nautobot_device_lifecycle_mgmt/inventoryitemsoftwarevalidationresult_list.html"
